requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
bcrypt==4.1.2
Werkzeug==3.0.1
email-validator==2.1.0
PyMySQL==1.1.0
mysql-connector-python==8.2.0
cryptography==41.0.8
